+++ package.json
@@ -95,7 +95,6 @@
     "@vitejs/plugin-vue-jsx": "^1.3.10",
     "@vitest/ui": "^2.1.8",
     "@vue/babel-plugin-jsx": "1.2.2",
-    "@vue/compiler-sfc": "^3.3.4",
     "@vue/eslint-config-typescript": "^10.0.0",
     "@vue/test-utils": "^2.4.1",
     "autoprefixer": "^10.4.14",
@@ -120,14 +119,12 @@
     "husky": "^7.0.4",
     "ignore": "^5.2.4",
     "jsdom": "^19.0.0",
-    "less": "^4.1.3",
     "lint-staged": "^13.2.3",
     "mockdate": "^3.0.5",
     "msw": "^1.2.3",
     "npm-run-all2": "^7.0.2",
     "postcss": "^8.4.26",
     "prettier": "2.8.1",
-    "prismjs": "^1.29.0",
     "raf": "^3.4.1",
     "rimraf": "^5.0.1",
     "rollup": "^2.79.1",
@@ -142,19 +139,8 @@
     "rollup-plugin-styles": "^4.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-vue": "^6.0.0",
-    "semver": "^7.5.4",
-    "tdesign-icons-view": "^0.3.4",
     "tdesign-publish-cli": "^0.0.12",
-    "tdesign-site-components": "0.16.0-alpha.6",
-    "tdesign-theme-generator": "^1.0.11",
-    "typescript": "^5.7.3",
-    "vite": "^2.9.16",
-    "vite-plugin-pwa": "^0.12.8",
-    "vite-plugin-tdoc": "^2.0.4",
-    "vitest": "^2.1.8",
-    "vue": "3.3.9",
-    "vue-router": "^4.2.4",
-    "workbox-precaching": "^6.6.0"
+    "typescript": "^5.7.3"
   },
   "config": {
     "commitizen": {
