+++ packages/tdesign-vue-next/site/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@tdesign/site",
+  "name": "@tdesign/vue-next-site",
   "private": true,
   "scripts": {
     "start": "pnpm run dev",
@@ -9,5 +9,24 @@
     "preview": "vite build --mode preview"
   },
   "author": "tdesign",
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "semver": "^7.5.4",
+    "vue": "3.3.9",
+    "vue-router": "^4.2.4",
+    "workbox-precaching": "^6.6.0"
+  },
+  "devDependencies": {
+    "@vue/compiler-sfc": "^3.3.4",
+    "@vitejs/plugin-vue": "^2.3.4",
+    "@vitejs/plugin-vue-jsx": "^1.3.10",
+    "less": "^4.1.3",
+    "prismjs": "^1.29.0",
+    "vite": "^2.9.16",
+    "vite-plugin-pwa": "^0.12.8",
+    "vite-plugin-tdoc": "^2.0.4",
+    "tdesign-icons-view": "^0.3.4",
+    "tdesign-site-components": "0.16.0-alpha.6",
+    "tdesign-theme-generator": "^1.0.11"
+  }
 }
