+++ packages/tdesign-vue-next/site/plugin-doc/md-to-vue.js
@@ -2,7 +2,7 @@
 import fs from 'fs';
 import path from 'path';
 import matter from 'gray-matter';
-import { compileUsage, getGitTimestamp } from '../../packages/common/docs/compile';
+import { compileUsage, getGitTimestamp } from '../../../../packages/common/docs/compile';
 import camelCase from 'camelcase';
 
 import testCoverage from '../test-coverage';
@@ -196,7 +196,7 @@ async function customRender({ source, file, md }) {
       componentName,
       usage: pageData.usage,
       demoPath: path.posix
-        .resolve(__dirname, `../../packages/components/${componentName}/_usage/index.vue`)
+        .resolve(__dirname, `../../../../packages/components/${componentName}/_usage/index.vue`)
         .replace(/\\/g, '/'),
     });
     if (usageObj) {
@@ -223,7 +223,7 @@ async function customRender({ source, file, md }) {
 
   // 设计指南内容 不展示 design Tab 则不解析
   if (pageData.isComponent && pageData.tdDocTabs.some((item) => item.tab === 'design')) {
-    const designDocPath = path.resolve(__dirname, `../../packages/common/docs/web/design/${componentName}.md`);
+    const designDocPath = path.resolve(__dirname, `../../../../packages/common/docs/web/design/${componentName}.md`);
 
     if (fs.existsSync(designDocPath)) {
       const designDocLastUpdated =
