+++ packages/tdesign-vue-next/site/plugin-doc/transforms.js
@@ -18,8 +18,8 @@ export default {
 
     // 统一换成 common 公共文档内容
     if (fileName && source.includes(':: BASE_DOC ::')) {
-      const localeDocPath = path.resolve(__dirname, `../../packages/common/docs/web/api/${fileName}`);
-      const defaultDocPath = path.resolve(__dirname, `../../packages/common/docs/web/api/${componentName}.md`);
+      const localeDocPath = path.resolve(__dirname, `../../../../packages/common/docs/web/api/${fileName}`);
+      const defaultDocPath = path.resolve(__dirname, `../../../../packages/common/docs/web/api/${componentName}.md`);
       let baseDoc = '';
       if (fs.existsSync(localeDocPath)) {
         // 优先载入语言版本
