+++ packages/tdesign-vue-next/site/src/app.vue
@@ -6,8 +6,8 @@
 
 <script setup>
 import { onMounted, ref } from 'vue';
-import cnConfig from 'tdesign-vue-next/es/locale/zh_CN';
-import enConfig from 'tdesign-vue-next/es/locale/en_US';
+import cnConfig from '@tdesign/components/locale/zh_CN';
+import enConfig from '@tdesign/components/locale/en_US';
 import { getLang } from 'tdesign-site-components';
 
 const globalConfig = ref(cnConfig);
