+++ packages/tdesign-vue-next/site/src/components/components.jsx
@@ -3,7 +3,8 @@ import semver from 'semver';
 import siteConfig from '../../site.config';
 import { htmlContent, mainJsContent, styleContent, packageJSONContent } from './codeSandbox/content';
 
-import packageJson from '../../../packages/tdesign-vue-next/package.json';
+// TODO
+import packageJson from '../../../package.json';
 
 const { docs, enDocs } = JSON.parse(JSON.stringify(siteConfig).replace(/component:.+/g, ''));
