+++ packages/tdesign-vue-next/site/src/components/demo-page.vue
@@ -10,8 +10,9 @@
 </template>
 
 <script>
-const demoVueReq = import.meta.globEager('../../../packages/components/**/_example/*.vue');
-const demoJsxReq = import.meta.globEager('../../../packages/components/**/_example/*.jsx');
+// TODO: 这种是否能优化一下呀，就不用路径的方式，用 @tdesign/components
+const demoVueReq = import.meta.globEager('../../../../../packages/components/**/_example/*.vue');
+const demoJsxReq = import.meta.globEager('../../../../../packages/components/**/_example/*.jsx');
 
 const demoReq = { ...demoVueReq, ...demoJsxReq };
 const demoObject = {};
