+++ packages/tdesign-vue-next/site/vite.config.js
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 import { VitePWA } from 'vite-plugin-pwa';
 import tDocPlugin from './plugin-doc';
 import pwaConfig from './pwaConfig';
-import { resolveConfig, basePlugin } from '../script/vite.base.config';
+import { resolveConfig, basePlugin } from '../../../script/vite.base.config';
 
 const publicPathMap = {
   preview: '/',
